sorting parts returned by aggregator
